finished AddAppointment component, modals were separated into two different components, made some dummy clients to be added for a new appointment, need to connect this to Schedule which will be fun, pictures can be replaced with the placeholderavatar I imported, both ClientList and AddAppointment performs crud operations, just need this to reflect in the database
